Move TileResult to own file
